+++ pipelex/plugins/anthropic/anthropic_llm_worker.py
@@ -6,7 +6,6 @@ from typing_extensions import override
 
 from pipelex import log
 from pipelex.cogt.exceptions import LLMCompletionError, LLMEngineParameterError, SdkTypeError
-from pipelex.cogt.inference.inference_report_delegate import InferenceReportDelegate
 from pipelex.cogt.llm.llm_job import LLMJob
 from pipelex.cogt.llm.llm_job_func import llm_job_func
 from pipelex.cogt.llm.llm_models.llm_engine import LLMEngine
@@ -15,6 +14,7 @@ from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
 from pipelex.config import get_config
 from pipelex.plugins.anthropic.anthropic_factory import AnthropicFactory
+from pipelex.reporting.reporting_protocol import ReportingProtocol
 from pipelex.tools.typing.pydantic_utils import BaseModelTypeVar
 
 
@@ -24,13 +24,13 @@ class AnthropicLLMWorker(LLMWorkerAbstract):
         sdk_instance: Any,
         llm_engine: LLMEngine,
         structure_method: Optional[StructureMethod] = None,
-        report_delegate: Optional[InferenceReportDelegate] = None,
+        reporting_delegate: Optional[ReportingProtocol] = None,
     ):
         LLMWorkerAbstract.__init__(
             self,
             llm_engine=llm_engine,
             structure_method=structure_method,
-            report_delegate=report_delegate,
+            reporting_delegate=reporting_delegate,
         )
         self.default_max_tokens: int
         if default_max_tokens := llm_engine.llm_model.max_tokens: