+++ pipelex/plugins/openai/openai_llm_worker.py
@@ -8,7 +8,6 @@ from typing_extensions import override
 
 from pipelex import log
 from pipelex.cogt.exceptions import LLMCompletionError, LLMEngineParameterError, LLMModelNotFoundError, SdkTypeError
-from pipelex.cogt.inference.inference_report_delegate import InferenceReportDelegate
 from pipelex.cogt.llm.llm_job import LLMJob
 from pipelex.cogt.llm.llm_job_func import llm_job_func
 from pipelex.cogt.llm.llm_models.llm_engine import LLMEngine
@@ -16,6 +15,7 @@ from pipelex.cogt.llm.llm_models.llm_family import LLMFamily
 from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.llm.structured_output import StructureMethod
 from pipelex.plugins.openai.openai_factory import OpenAIFactory
+from pipelex.reporting.reporting_protocol import ReportingProtocol
 from pipelex.tools.typing.pydantic_utils import BaseModelTypeVar
 
 
@@ -25,9 +25,9 @@ class OpenAILLMWorker(LLMWorkerAbstract):
         sdk_instance: Any,
         llm_engine: LLMEngine,
         structure_method: Optional[StructureMethod],
-        report_delegate: Optional[InferenceReportDelegate] = None,
+        reporting_delegate: Optional[ReportingProtocol] = None,
     ):
-        super().__init__(llm_engine=llm_engine, structure_method=structure_method, report_delegate=report_delegate)
+        super().__init__(llm_engine=llm_engine, structure_method=structure_method, reporting_delegate=reporting_delegate)
 
         if not isinstance(sdk_instance, openai.AsyncOpenAI):
             raise SdkTypeError(