+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [v0.2.13] - 2025-06-06
+
+- Added Discord badge on the Readme. Join the community! -> https://discord.gg/SReshKQjWt
+- Added a client for the Pipelex API. Join the waitlist -> https://www.pipelex.com/signup
+- Removed the `run_pipe_code` function. Replaced by `execute_pipeline` in `pipelex.pipeline.execute`.
+- Added llm deck `llm_for_img_to_text`.
+- Renamed `InferenceReportManager` to `ReportingManager`: It can report more than Inference cost. Renamed `InferenceReportDelegate` to `ReportingProtocol`.
+- Added an injection of dependency for `ReportingManager`
+- pipelex cli: fixed some bugs
+
 ## [v0.2.12] - 2025-06-03
 
 - pipelex cli: Split `pipelex init` into 2 separate functions: `pipelex init-libraries` and `pipelex init-config`